+++ noodles-vcf/CHANGELOG.md
@@ -10,6 +10,8 @@
   * vcf/header/info: Add conversion from `vcf::header::Record` to
     `vcf::header::Info` with a file format constraint.
 
+  * vcf/record/genotypes/keys: Implement `Default`.
+
 ### Changed
 
   * vcf/async/reader: `Reader::{records, query}` now parse records with the